test(runtime): split restricted store regression module
